add purge feature

+++ WebCache.psm1
@@ -14,6 +14,8 @@ class WebCache {
         [WebCache]::HashFunc = New-Object System.Security.Cryptography.SHA256CryptoServiceProvider
 
         log "WebCache.Init: Expire set to: $([WebCache]::Expire.ToString('yyyy\/MM\/dd HH:mm'))"
+
+        [WebCache]::InitPurgeList()
     }
 
     static [string] GetContent([string]$url) {
@@ -35,6 +37,9 @@ class WebCache {
             if ($res.StatusCode -eq 200) {
                 $res.Content |Out-File -FilePath $fp -Encoding utf8
                 logv "WebCache.GetContent: Cache saved: $fp, $($res.Content.Length)"
+
+                [WebCache]::PurgeCacheFile()
+
                 return $res.Content -join("`n")
             } else {
                 logerror "WebCache.GetContent: Web access error $($res.StatusCode)"
@@ -54,12 +59,15 @@ class WebCache {
         $purgeDate = [DateTime]::Now.AddDays(-31)
         $files = Get-ChildItem -File -Path ([WebCache]::CacheDir) |? {$_.LastWriteTime -le $purgeDate } |Sort LastWriteTime
         [WebCache]::PurgeList = $files.FullName
+        log "WebCache: PurgeList created: count=$([WebCache]::PurgeList.Count)"
     }
 
     static PurgeCacheFile() {
-        $f = [WebCache]::PurgeList[0]
-        Remove-Item -Path $f -Force -Confirm:$false -ErrorAction SilentlyContinue
-        [WebCache]::PurgeList = [WebCache]::PurgeList -ne $f
-        logv "WebCache: Cache purged: $f"
+        if ([WebCache]::PurgeList.Count) {
+            $f = [WebCache]::PurgeList[0]
+            Remove-Item -Path $f -Force -Confirm:$false -ErrorAction SilentlyContinue
+            [WebCache]::PurgeList = [WebCache]::PurgeList -ne $f
+            logv "WebCache: Cache purged: $f"
+        }
     }
 }
